Texture redirects, Workflow system, Overlay renderer and other short sprint things

+++ core/src/main/java/com/draniksoft/ome/editor/support/container/MoveDesc.java
@@ -4,8 +4,8 @@ public class MoveDesc {
 
     public int s, e;
     public int x, y;
-    public Float sx;
-    public Float sy;
+    public int sx;
+    public int sy;
 
 
 }
